moved _create_s3_bucket back from conftest as imports don't work if eo-learn is not installed in editable mode
